Add files via upload
Week 5 updates, new dprGUI class created.

+++ D&D Better Balance/src/dpr/calc/PlayerCharacter.java
@@ -8,12 +8,12 @@ public class PlayerCharacter {
 	
 	public static String name = "Player Character";		//character name for reference
 	public static String characterClass = "Fighter";	//character class for reference
-	public static int attackBonus = 0;					//modifier to hit
-	public static int armorClass = 10;					//armor class of character
-	public static int targetArmorClass = 10;			//armor class of target that is being hit
+	public static double attackBonus = 0;					//modifier to hit
+	public static double armorClass = 10;					//armor class of character
+	public static double targetArmorClass = 10;			//armor class of target that is being hit
 	public static double avgDiceDmg = 3;				//average value of the damage dice without modifiers
-	public static int dmgModifier = 0;					//modifier to each damage roll
-	public static int attacksPerRound = 1;				//attacks per round
+	public static double dmgModifier = 0;					//modifier to each damage roll
+	public static double attacksPerRound = 1;				//attacks per round
 	public static double sneakAvgAttackDmg = 0;			//optional sneak attack
 	public static double extraCritDmg = 0;				//optional extra crit damage
 	public static boolean extraCritAttack = false;		//optional extra crit attack
